services/webserver/requirements.txt
uvicorn
fastapi
python-multipart
debugpy
bcrypt
pyjwt
pymongo
sshtunnel
python-dateutil
requests
omise
pytz
aiohttp
kafka-python3
python-socketio